+++ packages/mg-components/.storybook/preview.ts
@@ -22,11 +22,13 @@ export const parameters: Preview['parameters'] = {
   docs: {
     extractArgTypes,
     extractComponentDescription,
-    transformSource: (_, ctx) => getStoryHTML(ctx.originalStoryFn(ctx.args)),
+    source: {
+      transform: (_: string, ctx: any) => getStoryHTML(ctx.originalStoryFn(ctx.args)),
+    },
   },
   options: {
     storySort: {
-      order: ['Intro', 'Atoms', 'Molecules', 'Style'],
+      order: ['Atoms', 'Molecules', 'Style'],
     },
   },
 };