+++ packages/mg-components/src/components/atoms/mg-tag/docs/mg-tag.stories.tsx
@@ -1,6 +1,5 @@
 import { h } from '@stencil/core';
 import { filterArgs } from '@mgdis/stencil-helpers';
-import { variants } from '../mg-tag.conf';
 import type { MgTag as MgTagType } from '../mg-tag';
 
 export default {
@@ -13,14 +12,16 @@ export default {
  * @param args - component arguments
  * @returns HTMLElement
  */
-const Template = (args: MgTagType & { slot: string }): HTMLElement => <mg-tag {...filterArgs(args, { variant: variants[0] })} innerHTML={args.slot}></mg-tag>;
+const Template = (args: MgTagType & { '': string }): HTMLElement => <mg-tag {...filterArgs(args, { variant: 'primary' }, [''])} innerHTML={args['']}></mg-tag>;
 
 export const MgTag = {
   render: Template,
   args: {
-    slot: 'Label',
-    variant: variants[0],
-    outline: false,
-    soft: false,
+    // Props
+    'variant': undefined,
+    'outline': false,
+    'soft': false,
+    // Slot
+    '': 'Label',
   },
 };