+++ packages/mg-components/src/components/molecules/inputs/mg-input-numeric/docs/mg-input-numeric.stories.tsx
@@ -1,6 +1,5 @@
 import { h } from '@stencil/core';
 import { filterArgs } from '@mgdis/stencil-helpers';
-import { types } from '../mg-input-numeric.conf';
 import type { MgInputNumeric as MgInputNumericType } from '../mg-input-numeric';
 
 export default {
@@ -14,55 +13,61 @@ export default {
  * @param args - component arguments
  * @returns HTMLElement
  */
-const Template = (args: MgInputNumericType): HTMLElement => <mg-input-numeric {...filterArgs(args)}></mg-input-numeric>;
+const Template = (args: MgInputNumericType): HTMLElement => (
+  <mg-input-numeric
+    {...filterArgs(
+      args,
+      {
+        type: 'decimal',
+        currency: 'EUR',
+        decimalLength: 2,
+        format: 'number',
+        integerLength: 13,
+        unitDisplay: 'short',
+        tooltipPosition: 'input',
+      },
+      ['append-input'],
+    )}
+    innerHTML={args['append-input']}
+  ></mg-input-numeric>
+);
 
 export const MgInputNumeric = {
   render: Template,
   args: {
-    // Global
-    value: '',
-    identifier: 'identifier',
-    name: 'input-name',
-    type: types[0], // decimal
-    // Label
-    label: 'Label',
-    labelOnTop: false,
-    labelHide: false,
-    // Input
-    placeholder: 'placeholder',
-    required: true,
-    disabled: false,
-    readonly: false,
-    max: undefined,
-    min: undefined,
-    mgWidth: undefined,
-    // Tooltip
-    tooltip: 'This is a tooltip',
-    tooltipPosition: undefined,
-    // Help Text
-    helpText: 'Help text with html <b>bold</b>, <em>italic</em>.',
+    // Props
+    'value': '',
+    'identifier': 'identifier',
+    'name': 'input-name',
+    'label': 'Label',
+    'labelOnTop': false,
+    'labelHide': false,
+    'placeholder': 'placeholder',
+    'required': false,
+    'readonly': false,
+    'max': '',
+    'min': '',
+    'disabled': false,
+    'mgWidth': undefined,
+    'tooltip': 'This is a tooltip',
+    'tooltipPosition': undefined,
+    'helpText': 'Help text with html <b>bold</b>, <em>italic</em>.',
+    'type': undefined,
+    'format': undefined,
+    'currency': '',
+    'unit': '',
+    'unitDisplay': undefined,
+    'integerLength': undefined,
+    'decimalLength': undefined,
+    // Slot
+    'append-input': '',
   },
 };
 
-const TemplateSlot = args => {
-  const labelOnTop = args.labelOnTop;
-  delete args.labelOnTop;
-  const labelHide = args.labelHide;
-  delete args.labelHide;
-  const helpText = args.helpText;
-  delete args.helpText;
-  // return element
-  return (
-    <mg-input-numeric {...args} label-on-top={labelOnTop} label-hide={labelHide} help-text={helpText}>
-      <span slot="append-input">km</span>
-    </mg-input-numeric>
-  );
-};
-
 export const AppendSlot = {
-  render: TemplateSlot,
-
+  render: Template,
   args: {
     ...MgInputNumeric.args,
+    'append-input': `<span slot="append-input">km</span>`,
   },
 };