+++ packages/mg-components/src/components/molecules/inputs/mg-input-text/docs/mg-input-text.stories.tsx
@@ -13,27 +13,48 @@ export default {
  * @param args - component arguments
  * @returns HTMLElement
  */
-const Template = (args: MgInputTextType): HTMLElement => {
-  return <mg-input-text {...filterArgs(args)}></mg-input-text>;
-};
-
+const Template = (args: MgInputTextType): HTMLElement => (
+  <mg-input-text
+    {...filterArgs(
+      args,
+      {
+        type: 'text',
+        maxlength: 400,
+        mgWidth: 'full',
+        tooltipPosition: 'input',
+      },
+      ['append-input'],
+    )}
+    innerHTML={args['append-input']}
+  ></mg-input-text>
+);
 export const MgInputText = {
   render: Template,
   args: {
-    // Global
-    value: '',
-    identifier: 'identifier',
-    name: 'input-name',
-    // Label
-    label: 'Label',
-    labelOnTop: false,
-    labelHide: false,
-    // Input
-    placeholder: 'placeholder',
-    // Tooltip
-    tooltip: 'This is a tooltip',
-    // Help Text
-    helpText: 'Help text with html <b>bold</b>, <em>italic</em>.',
+    // Props
+    'value': '',
+    'identifier': 'identifier',
+    'name': 'input-name',
+    'label': 'Label',
+    'labelOnTop': false,
+    'labelHide': false,
+    'type': undefined,
+    'icon': undefined,
+    'placeholder': 'placeholder',
+    'datalistoptions': undefined,
+    'maxlength': undefined,
+    'required': false,
+    'readonly': false,
+    'disabled': false,
+    'mgWidth': undefined,
+    'pattern': undefined,
+    'patternErrorMessage': undefined,
+    'tooltip': 'This is a tooltip',
+    'tooltipPosition': undefined,
+    'characterLeftHide': false,
+    'helpText': 'Help text with html <b>bold</b>, <em>italic</em>.',
+    // Slot
+    'append-input': '',
   },
 };
 
@@ -41,12 +62,11 @@ export const Type = {
   render: Template,
   args: {
     ...MgInputText.args,
-    // remove feature to focus on pattern
+    type: 'url',
+    // remove extra
     tooltip: undefined,
     label: 'Site web',
     helpText: undefined,
-    // Add pattern Email rules
-    type: 'url',
   },
 };
 
@@ -54,53 +74,31 @@ export const Pattern = {
   render: Template,
   args: {
     ...Type.args,
-    // remove feature to focus on pattern
-    tooltip: undefined,
-    label: 'Adresse email',
-    helpText: undefined,
-    // Add pattern Email rules
     type: 'email',
+    label: 'Adresse email',
     pattern:
       /^[a-zA-Z0-9!#$%&'*+\/=?^_`\{\|\}~\-]+(?:\.[a-zA-Z0-9!#$%&'*+\/=?^_`\{\|\}~\-]+)*@(?:[a-zA-Z0-9](?:[a-zA-Z0-9\-]*[a-zA-Z0-9])?\.)+[a-zA-Z0-9](?:[a-zA-Z0-9\-]*[a-zA-Z0-9])?$/
         .source, // From https://gitlab.mgdis.fr/core/core-back/core/-/blob/master/packages/validators/src/email/email.ts#L10
   },
 };
 
-/**
- * Template
- * @param args - component arguments
- * @returns HTMLElement
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-const SearchTemplate = (args: MgInputTextType): HTMLElement => {
-  // return element
-  return (
-    <form role="search">
-      <mg-input-text {...filterArgs(args)}>
-        <mg-button slot="append-input" label="search">
-          <mg-icon icon="magnifying-glass"></mg-icon> Search
-        </mg-button>
-      </mg-input-text>
-    </form>
-  );
-};
-
 export const Search = {
-  render: SearchTemplate,
-
+  render: Template,
   args: {
     ...MgInputText.args,
-    type: 'search',
-    icon: 'magnifying-glass',
+    'type': 'search',
+    'icon': 'magnifying-glass',
+    // Slot
+    'append-input': `<mg-button slot="append-input" label="search">
+        <mg-icon icon="magnifying-glass"></mg-icon>Search
+      </mg-button>`,
   },
 };
 
 export const Datalist = {
   render: Template,
-
   args: {
     ...MgInputText.args,
-    type: 'text',
     icon: 'magnifying-glass',
     datalistoptions: ['agent', 'admin', 'user'],
   },
@@ -110,9 +108,7 @@ export const DatalistOption = {
   render: Template,
 
   args: {
-    ...MgInputText.args,
-    type: 'text',
-    icon: 'magnifying-glass',
+    ...Datalist.args,
     datalistoptions: [
       { title: 'agent', value: '/agent/123' },
       { title: 'admin', value: '/admin/123' },