+++ packages/mg-components/src/components/molecules/inputs/mg-input-toggle/mg-input-toggle.tsx
@@ -92,7 +92,7 @@ export class MgInputToggle {
   /**
    * Define if label is displayed on top
    */
-  @Prop() labelOnTop?: boolean;
+  @Prop() labelOnTop = false;
 
   /**
    * Define if label is visible