+++ packages/mg-components/src/components/molecules/mg-panel/docs/mg-panel.stories.tsx
@@ -1,6 +1,5 @@
 import { h } from '@stencil/core';
 import { filterArgs } from '@mgdis/stencil-helpers';
-import { expandToggleDisplays, titlePositions } from '../mg-panel.conf';
 import type { MgPanel as MgPanelType } from '../mg-panel';
 
 export default {
@@ -15,28 +14,42 @@ export default {
  * @returns HTMLElement
  */
 const Template = (args: MgPanelType): HTMLElement => (
-  <mg-panel {...filterArgs(args)}>
-    <div>Content</div>
-    <div slot="header-right" style={{ width: '100%' }}>
-      <mg-badge label="1" value="1" style={{ 'margin-right': 'auto' }}></mg-badge>
-      <mg-button variant="secondary">
-        <mg-icon icon="file-upload"></mg-icon> Upload
-      </mg-button>
-      <mg-button is-icon variant="secondary" label="delete">
-        <mg-icon icon="trash"></mg-icon>
-      </mg-button>
-    </div>
-  </mg-panel>
+  <mg-panel
+    {...filterArgs(
+      args,
+      {
+        titlePosition: 'left',
+        expandToggleDisplay: 'text',
+      },
+      ['', 'header-right'],
+    )}
+    innerHTML={`${args['']}${args['header-right']}`}
+  ></mg-panel>
 );
 
 export const MgPanel = {
   render: Template,
   args: {
-    panelTitle: 'title',
-    expanded: false,
-    titleEditable: true,
-    expandToggleDisplay: expandToggleDisplays[0],
-    titlePosition: titlePositions[0],
-    expandToggleDisabled: false,
+    // Props
+    'identifier': undefined,
+    'panelTitle': 'title',
+    'titlePattern': undefined,
+    'titlePatternErrorMessage': undefined,
+    'titleEditable': false,
+    'titlePosition': undefined,
+    'expanded': false,
+    'expandToggleDisplay': undefined,
+    'expandToggleDisabled': false,
+    // Slots
+    '': `<div>Content</div>`,
+    'header-right': `<div slot="header-right" style="width:100%">
+  <mg-badge label="item" value="1" style="margin-right:auto"></mg-badge>
+  <mg-button variant="secondary">
+    <mg-icon icon="file-upload"></mg-icon> Upload
+  </mg-button>
+  <mg-button is-icon variant="secondary" label="delete">
+    <mg-icon icon="trash"></mg-icon>
+  </mg-button>
+</div>`,
   },
 };